api/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
flask-restx==1.1.0
pymongo==4.5.0
flask_pymongo==2.3.0
pytz==2022.1
gunicorn==21.2.0
deepdiff==6.7.1
marshmallow==3.21.2
werkzeug==3.0.1
user-agents==2.2.0
cachetools==5.3.3
python-dotenv==1.0.1
requests==2.32.3
ijson==3.3.0
typing-extensions==4.12.2
openai==1.68.2
